+++ source/blood/src/pqueue.h
@@ -30,14 +30,14 @@ public:
     void Upheap(void);
     void Downheap(unsigned int);
     void Delete(unsigned int);
-    void Insert(unsigned long, unsigned long);
-    unsigned long Remove(void);
+    void Insert(unsigned int, unsigned int);
+    unsigned int Remove(void);
 
     struct queueItem
     {
-        unsigned long at0; // priority
-        unsigned long at4; // data
+        unsigned int at0; // priority
+        unsigned int at4; // data
     } queueItems[kPQueueSize+1];
 
-    unsigned long fNodeCount; // at2008
+    unsigned int fNodeCount; // at2008
 };
